requirements.txt
requests==2.32.4
streamlit==1.46.1
pandas==2.3.0
pydub==0.25.1
audio-recorder-streamlit==0.0.10
speechbrain==1.0.3
python-dotenv==1.1.1
assemblyai==0.41.5
torch==2.7.1
torchaudio==2.7.1
soundfile==0.13.1
pydantic_settings-2.10.1